VT/requirements.txt
nest_asyncio>=1.5.1
python-dotenv>=0.19.0
requests>=2.26.0
pandas>=1.3.3
matplotlib>=3.4.3
tqdm>=4.62.3
pyarrow>=5.0.0
PyPDF2>=1.26.0
vt-py==0.17.5
requests==2.31.0
